Add files "Downloads PDF files from URLs.py" and "scrap PDF from Asif.py"
I divided the task into two subtasks.
A. Find the URL of all the PDF files of the articles. Using the scrapy library I built a bot (spider) that runs over the entire site and saves every URL of an article to a file.
The file URL_ok.txt contains the 12,376 files found
Executed in the file "Scrap PDF from Asif.py"
B. Download the files from the site and save them. Executed in the file "Downloads PDF files from URLs.py"